:books: [Docs]: Update LoadingItem.tsx
- Create directory of LoadingItem
- Add file of style.ts in LoadingItem directory
